+++ src/main/java/com/example/spring_crud/dto/notice/NoticeEditRequestDto.java
@@ -1,0 +1,13 @@
+package com.example.spring_crud.dto.notice;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class NoticeEditRequestDto {
+    private String title;
+    private String content;
+}
