+++ src/main/java/com/example/spring_crud/dto/notice/SearchNoticesDateRequestDto.java
@@ -1,0 +1,18 @@
+package com.example.spring_crud.dto.notice;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDate;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SearchNoticesDateRequestDto {
+    @DateTimeFormat(pattern = "yyyy-mm-dd")
+    private LocalDate startDate;
+    @DateTimeFormat(pattern = "yyyy-mm-dd")
+    private LocalDate endDate;
+}
