+++ src/main/java/com/example/spring_crud/dto/notice/SearchNoticesTitleRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.spring_crud.dto;
+package com.example.spring_crud.dto.notice;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,11 +9,7 @@ import javax.validation.constraints.NotBlank;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class NoticeCreateRequestDto {
+public class SearchNoticesTitleRequestDto {
     @NotBlank(message = "게시글 제목을 입력해주세요.")
     private String title;
-    @NotBlank(message = "게시글 내용을 입력해주세요.")
-    private String content;
-    @NotBlank(message = "게시글 작성자를 입력해주세요.")
-    private String writer;
 }
